+++ lib/enum.ex
@@ -1196,7 +1196,7 @@ defmodule Aja.Enum do
     size = RawVector.size(vector)
 
     if amount < size do
-      RawVector.slice(vector, 0, amount - 1)
+      RawVector.slice(vector, 0, amount - 1, 1)
     else
       RawVector.to_list(vector)
     end
@@ -1207,7 +1207,7 @@ defmodule Aja.Enum do
     start = amount + size
 
     if start > 0 do
-      RawVector.slice(vector, start, size - 1)
+      RawVector.slice(vector, start, size - 1, 1)
     else
       RawVector.to_list(vector)
     end
@@ -1233,7 +1233,7 @@ defmodule Aja.Enum do
     size = RawVector.size(vector)
 
     if amount < size do
-      RawVector.slice(vector, amount, size - 1)
+      RawVector.slice(vector, amount, size - 1, 1)
     else
       []
     end
@@ -1244,7 +1244,7 @@ defmodule Aja.Enum do
     last = amount + size
 
     if last > 0 do
-      RawVector.slice(vector, 0, last - 1)
+      RawVector.slice(vector, 0, last - 1, 1)
     else
       []
     end